==> Markdown/4. ExtractByEarliestReportingDate.Rmd <==
---
title: "4. ExtractByEarliestReportingDate"
output: word_document
---

```{r}
setwd("Z:\\Public\\BOM_Stationdata")

# find oldest weather stations in Australia
stnFiles <- dir("Historical_data",pattern="StnDet")
maxyear <-  1990 # no stations that started reporting after 1900

#Station Data
for (i in 1:length(stnFiles)) {
  stndata <- read.csv(paste("historical_data/",stnFiles[i],sep=""), stringsAsFactors=FALSE)
  olddata <- stndata[stndata[,14]<maxyear,]
  if (i == 1) {Allold <- olddata} else Allold <- rbind(Allold,olddata)
}

# plot the stations in space
require(ggplot2)
p <- ggplot(Allold,aes(x=Longitude.to.4.decimal.places.in.decimal.degrees,
                       y=Latitude.to.4.decimal.places.in.decimal.degrees)) +
  geom_point(aes(col=First.year.of.data.supplied.in.data.file))
p

#now subset by least missing data - Using >= 90% complete and >= 90% Quality Y as an arbitrary benchmark
station_quality <- Allold[which(as.numeric(Allold$Percentage.complete.between.first.and.last.records) >= 90 &
                                  as.numeric(Allold$Percentage.of.values.with.quality.flag..Y.) >= 90 &
                            as.numeric(Allold[,15]) >= 2010),]
```

Roundabout process to get a list of the files from the Z: drive. 

```{r}
############----------READING STREAMFLOW Rivers---------##############
if(dir.exists('Data/Rivers') == FALSE) {
  dir.create('Data/Rivers', recursive = TRUE)  
}

#name <- c("ID", "Station_Number", "Hydrstra_Site.ID", "LAT", "LON", "Site_Name", "Jurisdiction")
Rivers <- read.csv("Data/hrs_station_details.csv", skip = 11)
Rivers$AWRC.Station.Number <- as.character(Rivers$AWRC.Station.Number)
Rivers$Station.Name <- as.character(Rivers$Station.Name)
```

Create a Rivers folder. Extracting the HRS from the csv file and convert them to characters.

```{r}
setwd("F:\\University\\Honours\\Research Project\\Rainfall")
dir()

Rivers[32,4] <- -29.218 ##Changing Lat at Beardy River at Haystack because it was at 312 which is wrong

river.quality <- read.csv("Catchment Quality.csv", header = T)
Rivers <- merge(Rivers, river.quality, by = "Station_Number")
```

Merge 2 files to match the number of catchments - 111 here. 

```{r}
#######--------------FINDING NEAREST LAT/LON-----------################
#install.packages("rgeos")
library(rgeos)
library(sp)
Riverssp <- SpatialPoints(Rivers[,5:4])
Stationssp <- SpatialPoints(station_quality[,8:7])
Rivers$nearest_in_Stations <- apply(gDistance(Stationssp, Riverssp, byid=TRUE), 1, which.min)

Rivers <- data.frame(Rivers,  Stationssp[Rivers$nearest_in_Stations],station_quality[Rivers$nearest_in_Stations,][c(2,4)])
colnames(Rivers)[11:14] <- c("Nearest Lon", "Nearest Lat", "Station_No", "Station_Name")
```

Use the euclidean distance formula to find the station that is the minumum distance from the rain gauge. 

```{r}
Rivers$Difference <- sqrt((Rivers$LAT - Rivers$`Nearest Lat`)^2 + (Rivers$LON - Rivers$`Nearest Lon`)^2)
Rivers <- Rivers[order(Rivers$Difference, Rivers$Station_Number),]
Rivers <- Rivers[which(Rivers$Difference<0.45),]

#Write.csv in R
Abbreviation <- substr(Rivers$Site_Name, 1, 4)
Stations <- data.frame(Abbreviation, Rivers[,c(1,5,4)])

write.csv(Stations, file = "Final_Stations.csv", row.names = FALSE)
write.csv(Rivers[,c(1,6,4,5,7,8)], file = "Final_Rivers.csv", row.names = FALSE)
write.csv(Rivers, file = "Final_RiverStations.csv", row.names = FALSE)
```

Set an arbitrary cut-off value of 0.45 units for the values, then saved it in the files listed. 
Final_RiverStations.csv is the one used that will be used most frequently.
