==> Markdown/10. Observed summary Analysis.Rmd <==
---
title: "Raw data plots observed"
author: "Bruce"
date: "10 October 2017"
output:
  word_document: default
  pdf_document: default
---

```{r}
obs_1990 <- read.csv("obs.summary2_1990.csv", header = TRUE)
obs_2000 <- read.csv("obs.summary2_2000.csv", header = TRUE)
obs_2000$Number <- as.character(obs_2000$Number)
obs_1990$Number <- as.character(obs_1990$Number)

KG <- raster("koppen_ascii.txt")
plot(KG)

newproj <- "+proj=longlat +datum=WGS84"
coord_1990 <- cbind(obs_1990$LON, obs_1990$LAT)
pts_1990 <- SpatialPoints(coord_1990, proj4string=CRS(newproj))

coord_2000 <- cbind(obs_2000$LON, obs_2000$LAT)
pts_2000 <- SpatialPoints(coord_2000, proj4string = CRS(newproj))

obs_1990$KGext <- raster::extract(KG, pts_1990)
obs_2000$KGext <- raster::extract(KG, pts_2000)

obs_1990 <- obs_1990[c(1:104, 106),] ##Marambro SA has a stupid value in both 1990 and 2000. Too much leverage
obs_2000 <- obs_2000[c(1:100, 102:103),]
```

Creating a new column for Koppengeiger numbers - this was a suggestion from Floris to determine whether there were any links between the results and climate variations around AUstralia. 

```{r}
setwd("G:\\University\\Honours\\Data Access\\Modelling\\Results")

obs_1990 <- read.csv("obs.summary_1990.csv", header = TRUE)
obs_2000 <- read.csv("obs.summary_2000.csv", header = TRUE)

##PCA of observed
obs1_1990 <- obs_1990[,c(10:11,17:22)]
obs1_2000 <- obs_2000[,c(10:11,17:22)]
```

Taking all variables bar Number, Lat/Lon, Jurisdiction, NSE and r.sq.sqrt for the pca.

```{r}
obs1_1990[,c(1:2, 8)] <- log(obs1_1990[,c(1:2, 8)])
obs1_1990$FDC.high <- log(obs1_1990$FDC.high)
obs1_1990$FDC.low <- log(obs1_1990$FDC.low+0.1)
obs1_1990$FDC.mid <- log(obs1_1990$FDC.mid+0.1)

obs1_2000[,c(1:2)] <- log(obs1_2000[,c(1:2)])
obs1_2000$Peaks <- log(obs1_2000$Peaks+0.1)
obs1_2000$FDC.high <- log(obs1_2000$FDC.high)
obs1_2000$FDC.low <- log(obs1_2000$FDC.low+0.1)
obs1_2000$FDC.mid <- log(obs1_2000$FDC.mid+0.1)
```

Log-transform them to reduce high leverage. 

```{r}
obs1_1990pc <- prcomp(obs1_1990, scale = T)
obs1_2000pc <- prcomp(obs1_2000, scale = T)

par(mfrow = c(1,1))
#pca
library(car)
library(ggfortify)
autoplot(obs1_1990pc, data = obs_1990, colour = "KGext", 
         loadings = TRUE, loadings.label = TRUE, loadings.label.size = 5)
autoplot(obs1_2000pc, data = obs_2000, colour = "KGext",
         loadings = TRUE, loadings.label = TRUE, loadings.label.size = 5)
autoplot(obs1_1990pc, data = obs_1990, colour = "Type", 
         loadings = TRUE, loadings.label = TRUE, loadings.label.size = 5)
autoplot(obs1_2000pc, data = obs_2000, colour = "Type",
         loadings = TRUE, loadings.label = TRUE, loadings.label.size = 5)
```

Autoplot function from ggfortify was used to produce the graphs I wanted. 

```{r}
###Regression time

obs1_1990$KGext <- obs_1990$KGext
obs1_2000$KGext <- obs_2000$KGext
obs1_1990$Type <- obs_1990$Type
obs1_2000$Type <- obs_2000$Type

cor(obs1_1990[,1:8], method = "pearson")
cor(obs1_1990[,1:8], method = "spearman")
```

There seems to be some form of relationship in FDC.low, FDC.mid, Peaks, slope

I compared both the pearson and spearman correlation based on this link: https://stats.stackexchange.com/questions/8071/how-to-choose-between-pearson-and-spearman-correlation Where one of the commenters said: if you have S > P (spearman > pearson), that means that you have a correlation that is monotonic but not linear. Since it is good to have linearity in statistics (it is easier) you can try to apply a transformation on yy (such a log).

This is what the plots below are showing.

```{r}
plot(obs_2000$FDC.high, obs_2000$objFun, xlab = "FDC.high", ylab = "Viney")
plot(log(obs_2000$FDC.high), obs_2000$objFun, xlab = "log(FDC.high)", ylab = "Viney")
plot(obs_2000$FDC.low, obs_2000$objFun, xlab = "FDC.low", ylab = "Viney")
plot(log(obs_2000$FDC.low), obs_2000$objFun, xlab = "log(FDC.low)", ylab = "Viney")
plot(obs_2000$x2, obs_2000$objFun, xlab = "x2", ylab = "Viney")
plot(log(obs_2000$x2), obs_2000$objFun, xlab = "log(x2)", ylab = "Viney")
plot(obs_2000$AC, obs_2000$objFun, xlab = "AC", ylab = "Viney")
plot(log(obs_2000$AC), obs_2000$objFun, xlab = "log(AC)", ylab = "Viney")
plot(obs_2000$Peaks, obs_2000$objFun, xlab = "Peaks", ylab = "Viney")
plot(log(obs_2000$Peaks), obs_2000$objFun, xlab = "log(Peaks)", ylab = "Viney")
```

```{r}
# lm_obs1_1990 <- lm(objFun ~ KGext+x3+x4+Area+rbias+r.sq.log
#                    +AC, data = obs1_1990)
lm_obs1_1990 <- lm(Viney ~ ., data = obs1_1990)
#lm_obs1_1990 <- lm(Viney ~ .-CF-FDC.low-Slope, data = obs1_1990)
lm_obs1_1990 <- lm(Viney ~ Area + FDC.mid + Peaks, data = obs1_1990)
lm_obs1_1990 <- step(lm_obs1_1990, method = "both")
vif(lm_obs1_1990)
summary(lm_obs1_1990)
autoplot(lm_obs1_1990)
anova(lm_obs1_1990)

##Here, many of the variables are very highly correlated, such as the signatures 
#(FDC.low, mid, high, Peaks and slope) > 0.8. Cum.flow is also high correlated to runoff (which makes sense)
#Area, runoff, Peaks are significant. 
```
Here, regression analysis of 1990-1996 period. Variance inflation factors were used to show the ratio of variance with multiple terms - higher the worse it is, generally. 

```{r}
cor(obs1_2000[,1:9], method = "pearson", use = "pairwise.complete.obs")
cor(obs1_2000[,1:9], method = "spearman", use = "pairwise.complete.obs")

lm_obs1_2000 <- lm(Viney ~ .-FDC.low-FDC.mid,data = obs1_2000)
lm_obs1_2000 <- lm(Viney ~ Type, data = obs1_2000)
lm_obs1_2000 <- step(lm_obs1_2000, method = "backward")
#library(car)
vif(lm_obs1_2000)
summary(lm_obs1_2000)
autoplot(lm_obs1_2000)
anova(lm_obs1_2000)

#In the 1990 linear regression, Area, runoff, FDC,high, AC, and Peaks are significant
#In the 2000 linear regression, Area, FDC.mid, Peaks are significant. 
```

This was also done for the 2000-2006 period.
